+++ routes/admin_jumio.go
@@ -7,8 +7,10 @@ import (
 	"fmt"
 	"github.com/deso-protocol/backend/utils"
 	"github.com/deso-protocol/core/lib"
+	"github.com/golang/glog"
 	"io"
 	"net/http"
+	"strings"
 )
 
 type AdminResetJumioRequest struct {
@@ -304,5 +306,46 @@ func (fes *APIServer) AdminGetAllCountryLevelSignUpBonuses(ww http.ResponseWrite
 	if err := json.NewEncoder(ww).Encode(res); err != nil {
 		_AddBadRequestError(ww, fmt.Sprintf("GetAllCountryLevelSignUpBonuses: Encode failed: %v", err))
 		return
+	}
+}
+
+type CountrySignUpBonusResponse struct {
+	CountryLevelSignUpBonus CountryLevelSignUpBonus
+	CountryCodeDetails      utils.CountryCodeDetails
+}
+
+// SetAllCountrySignUpBonusMetadata goes through all countries in map of CountryCodes in utils and sets the sign-up
+// bonus config.
+func (fes *APIServer) SetAllCountrySignUpBonusMetadata() {
+	for countryCode, countryDetails := range utils.CountryCodes {
+		signUpBonus, err := fes.GetJumioCountrySignUpBonus(countryCode)
+		if err != nil {
+			glog.Errorf("SetAllCountrySignUpBonusMetadata: %v", err)
+			// If there was an error, look up the current value in the map and use that.
+			fes.GetSingleCountrySignUpBonus(countryCode)
+		}
+		fes.SetSingleCountrySignUpBonus(countryDetails, signUpBonus)
+	}
+}
+
+// SetSingleCountrySignUpBonus sets the sign up bonus configuration for a given country in the cached map.
+func (fes *APIServer) SetSingleCountrySignUpBonus(countryDetails utils.CountryCodeDetails,
+	signUpBonus CountryLevelSignUpBonus) {
+	fes.AllCountryLevelSignUpBonuses[countryDetails.Name] = CountrySignUpBonusResponse{
+		CountryLevelSignUpBonus: signUpBonus,
+		CountryCodeDetails:      countryDetails,
+	}
+}
+
+// GetSingleCountrySignUpBonus returns the current value of the sign-up bonus configuration stored in the cached map.
+func (fes *APIServer) GetSingleCountrySignUpBonus(countryCode string) CountryLevelSignUpBonus {
+	// Convert country code to uppercase just in case.
+	countryCodeDetails := utils.CountryCodes[strings.ToUpper(countryCode)]
+	// If we can't find the signup bonus from the map, return the default. Else, return the sign up bonus we found in
+	// the map.
+	if countrySignUpBonusResponse, exists := fes.AllCountryLevelSignUpBonuses[countryCodeDetails.Name]; !exists {
+		return fes.GetDefaultJumioCountrySignUpBonus()
+	} else {
+		return countrySignUpBonusResponse.CountryLevelSignUpBonus
 	}
 }
